replace old Process class with new
